chore: merge dom.ts into query.ts
